Minor Change in stich
Remaining ssues I'm aware of:
(1) File confusion whether using csv or txt!!
(2) Dempster CROPS am inital or final CLOSED state, so need to manually correct this in Excel etc as it stands.
(3) Chances of perfect match up with original are slim and so probably need to import that and column crop the length to the minimum of the three... or still, essentially match it up in Excel.