Had to add a bugfix for Sam gen dataNumanNoGPU
Some of the parquets (one channel data) was read with Channels as object not integer so bug fix for this.  May mean application quite specific to these data now, but used try/except so hopefully still runs with other formats. Really needs an option to say what to plot against what!!